@@ -1,6 +1,6 @@
 {
   "name": "@jridgewell/trace-mapping",
-  "version": "0.3.9",
+  "version": "0.3.17",
   "description": "Trace the original position through a source map",
   "keywords": [
     "source",
@@ -13,11 +13,15 @@
     "dist"
   ],
   "exports": {
-    ".": {
-      "browser": "./dist/trace-mapping.umd.js",
-      "require": "./dist/trace-mapping.umd.js",
-      "import": "./dist/trace-mapping.mjs"
-    },
+    ".": [
+      {
+        "types": "./dist/types/trace-mapping.d.ts",
+        "browser": "./dist/trace-mapping.umd.js",
+        "require": "./dist/trace-mapping.umd.js",
+        "import": "./dist/trace-mapping.mjs"
+      },
+      "./dist/trace-mapping.umd.js"
+    ],
     "./package.json": "./package.json"
   },
   "author": "Justin Ridgewell <justin@ridgewell.name>",
@@ -29,7 +33,7 @@
   "scripts": {
     "benchmark": "run-s build:rollup benchmark:*",
     "benchmark:install": "cd benchmark && npm install",
-    "benchmark:only": "node benchmark/index.mjs",
+    "benchmark:only": "node --expose-gc benchmark/index.mjs",
     "build": "run-s -n build:*",
     "build:rollup": "rollup -c rollup.config.js",
     "build:ts": "tsc --project tsconfig.build.json",
@@ -48,23 +52,24 @@
     "test:watch": "ava --watch"
   },
   "devDependencies": {
-    "@rollup/plugin-typescript": "8.3.0",
-    "@typescript-eslint/eslint-plugin": "5.10.0",
-    "@typescript-eslint/parser": "5.10.0",
-    "ava": "4.0.1",
+    "@rollup/plugin-typescript": "8.5.0",
+    "@typescript-eslint/eslint-plugin": "5.39.0",
+    "@typescript-eslint/parser": "5.39.0",
+    "ava": "4.3.3",
     "benchmark": "2.1.4",
-    "c8": "7.11.0",
-    "esbuild": "0.14.14",
-    "esbuild-node-loader": "0.6.4",
-    "eslint": "8.7.0",
-    "eslint-config-prettier": "8.3.0",
+    "c8": "7.12.0",
+    "esbuild": "0.15.10",
+    "eslint": "8.25.0",
+    "eslint-config-prettier": "8.5.0",
+    "eslint-plugin-no-only-tests": "3.0.0",
     "npm-run-all": "4.1.5",
-    "prettier": "2.5.1",
-    "rollup": "2.64.0",
-    "typescript": "4.5.4"
+    "prettier": "2.7.1",
+    "rollup": "2.79.1",
+    "tsx": "3.10.1",
+    "typescript": "4.8.4"
   },
   "dependencies": {
-    "@jridgewell/resolve-uri": "^3.0.3",
-    "@jridgewell/sourcemap-codec": "^1.4.10"
+    "@jridgewell/resolve-uri": "3.1.0",
+    "@jridgewell/sourcemap-codec": "1.4.14"
   }
 }
