@@ -4,6 +4,8 @@ import {
 } from "discord.js";
 
 import { Command } from "../../common/types/Command";
+import { postData, sendData } from "../../common/services/api";
+import { IUser } from "../../common/types/UserTypes";
 
 export default new Command({
   name: "glpi",
@@ -16,8 +18,29 @@ export default new Command({
       type: ApplicationCommandOptionType.String,
       required: true,
     },
+    {
+      name: "email",
+      description: "Digite seu e-mail",
+      type: ApplicationCommandOptionType.String,
+      required: true,
+    },
   ],
   run({ interaction, options }) {
-    interaction.reply({ ephemeral: true, content: "Vou sincronizar" });
+    const { user } = interaction;
+    const usernameGLPI = options.getString("nome");
+    const email = options.getString("email");
+
+    const body: IUser = {
+      usernameGLPI: usernameGLPI!,
+      email: email!,
+      discordId: user.id!,
+    };
+
+    sendData(body);
+
+    interaction.reply({
+      ephemeral: true,
+      content: `O usuário ${user} com id: ${user.id} solicitou: ${usernameGLPI}.`,
+    });
   },
 });
