@@ -10,8 +10,8 @@ const routes = Router();
 // routes.get("/users", userController.findAll);
 // routes.get("/users/:id", userController.find);
 routes.post("/users", userController.create);
-routes.delete("/users/:id", userController.delete);
-routes.put("/users/:id", userController.update);
+// routes.delete("/users/:id", userController.delete);
+// routes.put("/users/:id", userController.update);
 // routes.post("/products", productController.create);
 // routes.get("/products", productController.list);
 // routes.get("/products-stock", productController.getStock);
